+++ src/components/elements/VerifyCard.vue
@@ -19,45 +19,17 @@
             </md-card-content>
             <md-card-actions v-if="image !== ''" class="actions">
                 <md-button class="vote-button md-raised" @click="verifyPrompt = true">VOTE</md-button>
-                <md-button class="unct-button md-raised" @click="uncertainPrompt = true">UNCERTAIN</md-button>
-                <md-button class="omit-button md-raised" @click="omitPrompt = true">NO RISK</md-button>
             </md-card-actions>
         </md-card>
         <md-dialog :md-active="verifyPrompt">
-            <md-dialog-title class="dialog-title">Confirm Your Choice</md-dialog-title>
-            <md-dialog-content class="dialog-content">
+            <md-dialog-title class="dialog-title dialog-title-custom">Confirm Your Choice</md-dialog-title>
+            <md-dialog-content class="dialog-content dialog-content-custom">
                 <md-empty-state
                         md-icon="check_circle_outline"
                         md-description="Are you sure you wish to vote for this annotation? You may only pick one.">
                     <span class="button-span">
                         <md-button class="vote-button md-raised" @click="verify()">Confirm</md-button>
                         <md-button class="omit-button md-raised" @click="verifyPrompt = false">CANCEL</md-button>
-                    </span>
-                </md-empty-state>
-            </md-dialog-content>
-        </md-dialog>
-        <md-dialog :md-active="uncertainPrompt">
-            <md-dialog-title class="dialog-title">I don't know.</md-dialog-title>
-            <md-dialog-content class="dialog-content">
-                <md-empty-state
-                        md-icon="help_outline"
-                        md-description="Only use this options if you really do not know whether this object is potentially risky.">
-                    <span class="button-span">
-                        <md-button class="vote-button md-raised" @click="uncertain()">Confirm</md-button>
-                        <md-button class="omit-button md-raised" @click="uncertainPrompt = false">CANCEL</md-button>
-                    </span>
-                </md-empty-state>
-            </md-dialog-content>
-        </md-dialog>
-        <md-dialog :md-active="omitPrompt">
-            <md-dialog-title class="dialog-title">Is this object not risky?</md-dialog-title>
-            <md-dialog-content class="dialog-content">
-                <md-empty-state
-                        md-icon="remove_circle_outline"
-                        md-description="If you mark this object as not risky, you will not be able to vote on the other snapshots.">
-                    <span class="button-span">
-                        <md-button class="vote-button md-raised" @click="omit()">Confirm</md-button>
-                        <md-button class="omit-button md-raised" @click="omitPrompt = false">CANCEL</md-button>
                     </span>
                 </md-empty-state>
             </md-dialog-content>
@@ -79,8 +51,6 @@
         data: function() {
           return {
               verifyPrompt: false,
-              uncertainPrompt: false,
-              omitPrompt: false,
               selectedAnnotation: null
           }
         },
@@ -91,14 +61,6 @@
             verify: function() {
                 this.updateVerifyVotes(this.title);
                 this.verifyPrompt = false;
-            },
-            uncertain: function () {
-                this.updateVerifyVotes("?");
-                this.uncertainPrompt = false;
-            },
-            omit: function () {
-                this.updateVerifyVotes(null);
-                this.omitPrompt = false;
             }
         }
     };
@@ -119,41 +81,7 @@
         height: 100%;
     }
 
-    .button-span {
-        width: 100%;
-    }
-
-    span .vote-button, span .omit-button {
-        margin-left: 4px;
-        margin-right: 4px;
-    }
-
-    .vote-button, .omit-button, .unct-button {
-        color: white !important;
-    }
-
-    .vote-button {
-        background-color: var(--forest-green) !important;
-    }
-
-    .unct-button {
-        background-color: #a1a1a1 !important;
-    }
-
-    .omit-button {
-        background-color: #d32f2f !important;
-    }
-
     .actions {
         justify-content: center !important;
-    }
-
-    .dialog-title {
-        text-align: center;
-        margin-bottom: 0;
-    }
-
-    .dialog-content {
-        padding-bottom: 0;
     }
 </style>
